refactor(daemon-cli): move restart health formatting to diagnostics module
Move formatRestartFailure and the still-starting message resolver out of
lifecycle.ts and share the success-path emit between still-starting and
normal restarts. lifecycle.ts was pushed past the max-lines budget by the
still-starting wait handling; the diagnostics module stays the home for
restart message rendering.
